+++ package/example/lib/pages/custom_controls.dart
@@ -14,7 +14,10 @@ class CustomControlsExamplePage extends StatefulWidget {
 class _CustomControlsExamplePageState extends State<CustomControlsExamplePage> {
   final _meeduPlayerController = MeeduPlayerController(
       controlsStyle: ControlsStyle.custom,
-      enabledOverlays: const EnabledOverlays(volume: false));
+      enabledOverlays: const EnabledOverlays(volume: false),
+
+      
+      );
 
   StreamSubscription? _playerEventSubs;